+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/PneuResource.java
@@ -17,7 +17,6 @@ import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
 import org.glassfish.jersey.media.multipart.FormDataParam;
 import org.jetbrains.annotations.NotNull;
 
-import javax.validation.Valid;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import java.io.InputStream;
@@ -58,7 +57,7 @@ public final class PneuResource {
     @Secured(permissions = Pilares.Frota.Pneu.ALTERAR)
     @Path("/retornar-descarte")
     public PneuRetornoDescarteSuccess retornarPneuDescarte(
-            @Valid @NotNull @Required final PneuRetornoDescarte pneuRetornoDescarte) {
+            @NotNull @Required final PneuRetornoDescarte pneuRetornoDescarte) {
         return service.retornarPneuDescarte(pneuRetornoDescarte);
     }